requirements.txt
fastapi
uvicorn
streamlit
requests
boto3
opensearch-py
requests-aws4auth
python-dotenv
prometheus_fastapi_instrumentator
ragas
datasets
langchain-aws